Moving helpers into login service
