+++ src/common/helpers.ts
@@ -1,5 +1,3 @@
-import { IDiscordAuthResponse } from "./interfaces";
-
 export const getStoreUrl = (id: string) => {
     return `https://www.microsoft.com/store/apps/${id}`;
 }
@@ -11,28 +9,6 @@ export const getGithubUrl = (id: string) => {
 export const getDiscordUrl = (id: string) => {
     return `https://www.discord.gg/${id}`;
 }
-
-const AuthData = {
-    Get: GetDiscordAuthData,
-    Set: SetDiscordAuthData,
-    Clear: ClearDiscordAuthData
-};
-
-export function GetDiscordAuthData() : IDiscordAuthResponse | undefined {
-    const authData = localStorage.getItem("discordAuthData");
-    if(!authData) return;
-    return JSON.parse(authData);
-}
-
-export function SetDiscordAuthData(data: IDiscordAuthResponse) {
-    if(!data) return;
-    localStorage.setItem("discordAuthData", JSON.stringify(data));
-}
-
-export function ClearDiscordAuthData() {
-    localStorage.removeItem("discordAuthData");
-}
-
 export default {
-    getDiscordUrl, getGithubUrl, getStoreUrl, AuthData
+    getDiscordUrl, getGithubUrl, getStoreUrl
 }
