+++ src/common/loginService.ts
@@ -1,0 +1,23 @@
+import { IDiscordAuthResponse } from "./interfaces";
+
+
+export const AuthData = {
+    Get: GetDiscordAuthData,
+    Set: SetDiscordAuthData,
+    Clear: ClearDiscordAuthData
+};
+
+export function GetDiscordAuthData() : IDiscordAuthResponse | undefined {
+    const authData = localStorage.getItem("discordAuthData");
+    if(!authData) return;
+    return JSON.parse(authData);
+}
+
+export function SetDiscordAuthData(data: IDiscordAuthResponse) {
+    if(!data) return;
+    localStorage.setItem("discordAuthData", JSON.stringify(data));
+}
+
+export function ClearDiscordAuthData() {
+    localStorage.removeItem("discordAuthData");
+}
